jobs/daily-push/requirements.txt
functions-framework==3.4.0
google-cloud-firestore==2.13.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
